requirements.txt
Django==5.0.6
mysqlclient==2.2.4
python-decouple==3.8
Pillow==10.3.0
django-widget-tweaks==1.5.0
whitenoise==6.6.0
